refactor: Move getAddress and getAddresses functions into Keyring class file
